// apps/perp/src/types/charting-library.d.ts
export declare type Nominal<T, Name extends string> = T & {
  /* eslint-disable-next-line jsdoc/require-jsdoc */ [Symbol.species]: Name;
};
// ...
export interface Bar {
  /** Bar time.
   * Amount of **milliseconds** since Unix epoch start in **UTC** timezone.
   * `time` for daily, weekly, and monthly bars is expected to be a trading day (not session start day) at 00:00 UTC.
   * The library adjusts time according to `session` from {@link LibrarySymbolInfo}.
   */
  time: number;
  /** Opening price */
  open: number;
  /** High price */
  high: number;
  /** Low price */
  low: number;
  /** Closing price */
  close: number;
  /** Trading Volume */
  volume?: number;
}
// ...
export interface TimeFrameItem {
  /** Defines the range to be set for the given resolution. It has the `<integer><y|m|d>` ( \d+(y|m|d) as Regex ) format. */
  text: string;
  /** Resolution to be set */
  resolution: ResolutionString;
  /** Optional text displayed in the pop-up menu. If not set, `title` wil be used instead. */
  description?: string;
  /** Optional text representing the time frame. If not set a generated string will be set based on `text` property. */
  title?: string;
}
// ...
export type LanguageCode =
  | "ar"
  | "zh"
  | "ca_ES"
  | "en"
  | "fr"
  | "de"
  | "he_IL"
  | "id_ID"
  | "it"
  | "ja"
  | "ko"
  | "pl"
  | "pt"
  | "ru"
  | "es"
  | "sv"
  | "th"
  | "tr"
  | "vi"
  | "ms_MY"
  | "zh_TW";
// ...
export type ResolutionString = Nominal<string, "ResolutionString">;
// ...
export type SubscribeBarsCallback = (bar: Bar) => void;